Move backend-neutral LoRA helpers out of megatron_utils
