+++ miles/rollout/generate_utils/generate_endpoint_utils.py
@@ -8,7 +8,7 @@ from typing import Any
 import numpy as np
 import pybase64
 
-from miles.backends.megatron_utils.lora_utils import LORA_ADAPTER_NAME, is_lora_enabled
+from miles.utils.lora import LORA_ADAPTER_NAME, is_lora_enabled
 from miles.utils.processing_utils import encode_image_for_rollout_engine
 from miles.utils.types import Sample
