+++ miles/utils/lora.py
@@ -1,0 +1,8 @@
+from argparse import Namespace
+
+LORA_ADAPTER_NAME = "miles_lora"
+
+
+def is_lora_enabled(args: Namespace) -> bool:
+    """Check if LoRA is enabled based on arguments."""
+    return getattr(args, "lora_rank", 0) > 0 or getattr(args, "lora_adapter_path", None) is not None
